- refactored the code to use includes relative to the root (src)

+++ Minecraft Clone/src/Level/World/WorldParameters.h
@@ -1,11 +1,11 @@
 #pragma once
 
 struct WorldParameters {
-    size_t horizontal_ticking_distance_ = 16;
-    size_t vertical_ticking_distance_ = 8;
+    int horizontal_ticking_distance_ = 16;
+    int vertical_ticking_distance_ = 8;
 
-    size_t spawn_chunk_horizontal_radius_ = 10;
-    size_t spawn_chunk_vertical_radius_ = 10;
+    int spawn_chunk_horizontal_radius_ = 10;
+    int spawn_chunk_vertical_radius_ = 10;
 
     float gravity_ = 10.f;
